[VM] Ensure that a newly allocated object is either new or added to remembered set if we want to eliminate write barriers to it
For all non-generic classes the AllocateObject runtime call would not
guarantee that the returned object is new or marked as remembered. If
such an object is subject to a store with an eliminated write-barrier,
we have incorrect remembered set, which causes flaky crashes.

This CL fixes the AllocateObject runtime call to always ensure the
object is new or added to the remembered set if write-barrier
elimination is enabled for those.

Furthermore the conditions for when the write barrier can be eliminated
is centralized in one place instead of duplicated (in runtime_entry.cc
and il.h) to avoid future divergence of the two.

Similarly the condition to when an array uses card marking is put into
one place.

Fixes https://github.com/dart-lang/sdk/issues/36011

Change-Id: I270f3618a267a35fa6ea63dcf4092488c0319328
Reviewed-on: https://dart-review.googlesource.com/c/95022
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Reviewed-by: Samir Jindel <sjindel@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>