Update pkg:html version to 0.13.4+1
There was a stray 0.13.3+3 tag committed two years ago, that some builders
had in their cache, so they were not picking up a new 0.13.3+3 tag created
days ago. No tag may ever be changed or reused in our dependencies.  The 0.13.4
tag was created against a repo with pubspec version 0.13.4-dev, which was also
wrong, so we needed a fixed, released 0.13.4+1 version.

Change-Id: I868db93df13d4a24ff30bb49fcb190151ab53036
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/95240
Reviewed-by: William Hesse <whesse@google.com>