FE Server shouldn't try to invalidate other file schemes
When trying to invalidate files when using a multi-
root file system, if a .dill file that is being
used to initialize an incremental compiler contains
both files with a custom marker scheme, as well as
an internal one like `org-dartlang-sdk`, it will
throw when it encounters the internal scheme.

This commit remedies the issue, but this is my
first CL to the SDK, I'm not sure if this is
the best approach.

Other approaches may be:

- Ignore dartlang internal file schemes only
- Catch errors from entity.exists instead of filtering

And I'm sure there are many others.

Change-Id: Idc7aba4b41eb5cf7f66db492f0ea6e26252779c5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/94680
Commit-Queue: Alexander Aprelev <aam@google.com>
Reviewed-by: Alexander Aprelev <aam@google.com>