Use subtype rather than assignability to disambiguate maps/sets.
This causes the implementation to match the spec.  Also, it's an
important distinction because the analyzer allows the user to change
the meaning of assignability checks using the analysis option
"implicit_casts: false".  We don't want this option to change the
meaning of set/map disambiguation.

Change-Id: I2a870209a31b4344c0568ecfe741bf731d9489a1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/96120
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>