New DomExceptions added to browsers.
These new exceptions have no legacy values.

DataCloneError - The object can not be cloned.
EncodingError - The encoding or decoding operation failed.
NotReadableError - The input/output read operation failed
UnknownError - The operation failed for an unknown transient reason (e.g. out of memory).
ConstraintError - A mutation operation in a transaction failed because a constraint was not satisfied.
DataError - Provided data is inadequate.
TransactionInactiveError - A request was placed against a transaction which is currently not active, or which is finished.
ReadOnlyError - The mutating operation was attempted in a "readonly" transaction.
VersionError  An attempt was made to open a database using a lower version than the existing version.
OperationError - The operation failed for an operation-specific reason.
NotAllowedError - The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission.

R=vsm@google.com

Change-Id: I1245bc788fd287d566f42184ebe914165afa95ac
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/97163
Reviewed-by: Vijay Menon <vsm@google.com>
Commit-Queue: Terry Lucas <terry@google.com>