[gardening] Fix arguments test for app_jit(k).
The fix for https://github.com/dart-lang/sdk/issues/35960 broke
on app_jit(k) runs, because these expect the arguments to be
passed when creating the jit snapshot. Pass dartOptions along to
computeCompilerArguments as well, so that the
AppJitCompilerConfiguration can add them in appropriately.

Change-Id: I8c7b5a3a1689943db4e6c3785ccaf8bfdb839c2d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/97516
Reviewed-by: Stevie Strickland <sstrickl@google.com>
Commit-Queue: Stevie Strickland <sstrickl@google.com>