Test bytecode benchmarks when built with bytecode support.
This change adds a commit queue test that the bytecode benchmarks continue
to work in order to avoid any future unintended breakage.

The new linux-x64-bytecode build is exactly the same as the linux-x64 build
except without the SIMDBC build.

This change also moves the build from the script out to the test matrix,
which let's Goma work correctly, speeding the builder considerably up.

Change-Id: Ifc275c55e405f8dcd7e98a8f15c651602a6dfc4f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/98326
Reviewed-by: Alexander Markov <alexmarkov@google.com>