rework dartfix NNBD to use FixCodeTask.processPackage
This generalizes the dartfix NNBD processing of packages so that
other fixes can process packages in a similar manner.
In addition, if at least one package to be migrated to NNBD
cannot be marked as NNBD or is not already NNBD,
then no additional NNBD processing is done.

This also addresses comments in
https://dart-review.googlesource.com/c/sdk/+/98429

Change-Id: I406d9d395a0ee8a85cc5b3d7be311a76e1e48bdb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/98524
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Dan Rubel <danrubel@google.com>