Rename Nullability enum to avoid confusion with spec.
The spec uses the terminology "nullable", "non-nullable", etc. to
refer to properties of types that differ subtly from their structure.
For example "Null" is considered nullable, even though it lacks a
trailing "?".  Since we were using the Nullability enum to denote a
purely structural property of types (whether they end in "?", "*", or
neither), this led to confusion.

So to avoid confusion, we rename the Nullability enum, and its
enumerated values, to clarify that it refers specifically to the
nullability suffixes and not to the property of being
nullable/non-nullable.

Change-Id: Ia34a500a3d2d724d30c66a410140ae11d6512f50
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/99275
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>