[fuchsia] Mark subprocess images as executable
Currently, the VMO returned by `fdio_get_vmo_clone` is executable,
despite being documented as being read-only.  We wish to correct
the implementation, but doing so would break anything that relies
on the implicit executability of that VMO.

To enable Fuchsia to make this transition, we should explicitly call
zx_vmo_replace_as_executable on the VMO holding the binary image we
wish to launch as a subprocess.

Change-Id: I040ed2409120a58a30e88ef34a01fe73434a7ea0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/99369
Reviewed-by: Zach Anderson <zra@google.com>
Commit-Queue: Zach Anderson <zra@google.com>
Auto-Submit: Drew Fisher <zarvox@google.com>