[dart/fuzzer] restrict stmt depth and length
Rationale:
We have consistent failures with DBC reaching
the 64K constant limit. Although a true problem,
as this stage, it is not so useful to keep
DartFuzz nightly runs red constantly for this.
This CL reduces the length a bit, so that we
hit this restriction less often.

https://github.com/dart-lang/sdk/issues/36646
https://github.com/dart-lang/sdk/issues/36658

Change-Id: Ief2c7293b0159b798ee0426751899dc7fffd4f5d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/100464
Reviewed-by: Samir Jindel <sjindel@google.com>
Commit-Queue: Aart Bik <ajcbik@google.com>